Revert modules renaming
